Made changes

+++ SG.Web/Controllers/ContentCreatorController.cs
@@ -25,7 +25,7 @@ namespace SG.Web.Controllers
         {
             return Ok(_repository.GetAll());
         }
-        [HttpGet("get-by-id")]
+        [HttpGet("get-by-id/{id}")]
         public IActionResult GetById(Guid id)
         {
             var result = _repository.GetById(id);
@@ -33,20 +33,35 @@ namespace SG.Web.Controllers
             {
                 return Ok(result);
             }
-            return NotFound("Not Found");
+            return NotFound($"The id {id} cannot be found ");
         }
-        [HttpPatch("update")]
-        public IActionResult Update(ContentCreatorModel model)
+        [HttpPatch("update/{id}")]
+        public IActionResult Update(ContentCreatorModel model, Guid id)
         {
+            var result = _repository.GetById(id);
+            if (result != null)
+            {
+                result.WorkEmail = String.IsNullOrEmpty(model.WorkEmail) ? result.WorkEmail : model.WorkEmail;
+                result.Department = String.IsNullOrEmpty(model.Department) ? result.Department : model.Department;
+                result.DateModified = DateTime.Now;
+                result.Role = string.IsNullOrEmpty(model.Role) ? result.Role : model.Role;
+                _repository.Update(result);
+                return Ok(result);
 
-            _repository.Update(model);
-            return Ok(model);
+
+            }
+            return NotFound($"The id {id} cannot be found ");
         }
-        [HttpDelete("delete")]
+        [HttpDelete("delete/{id}")]
         public IActionResult Delete(Guid id)
         {
-            _repository.Delete(id);
-            return Ok("Deleted");
+            var result = _repository.GetById(id);
+            if (result != null)
+            {
+                _repository.Delete(id);
+                return Ok($"Content Creator with id {id} cannot be found");
+            }
+            return NotFound($"The id {id} cannot be found ");
         }
     }
 }
