+++ SG.Web/Controllers/LearningMaterialController.cs
@@ -26,7 +26,7 @@ namespace SG.Web.Controllers
         {
             return Ok(_repository.GetAll());
         }
-        [HttpGet("get-by-id")]
+        [HttpGet("get-by-id/{id}")]
         public IActionResult GetById(Guid id)
         {
             var result = _repository.GetById(id);
@@ -36,18 +36,36 @@ namespace SG.Web.Controllers
             }
             return NotFound("Not Found");
         }
-        [HttpPatch("update")]
-        public IActionResult Update(LearningMaterialModel model)
+        [HttpPatch("update/{id}")]
+        public IActionResult Update(LearningMaterialModel model, Guid id)
         {
+            var result = _repository.GetById(id);
+            if (result == null)
+            {
+                result.WorkEmail = String.IsNullOrEmpty(model.WorkEmail) ? result.WorkEmail : model.WorkEmail;
+                result.DateModified = DateTime.Now;
 
-            _repository.Update(model);
-            return Ok(model);
+                result.InternId = String.IsNullOrEmpty(model.InternId.ToString()) ? result.InternId : model.InternId;
+                result.IsChecked = model.IsChecked;
+                result.UploadModelId = String.IsNullOrEmpty(model.UploadModelId.ToString()) ? result.UploadModelId : model.UploadModelId;
+                _repository.Update(result);
+                return Ok(result);
+
+            }
+
+            // _repository.Update(model);
+            return NotFound($"The id {id} cannot be found ");
         }
         [HttpDelete("delete")]
         public IActionResult Delete(Guid id)
         {
-            _repository.Delete(id);
-            return Ok("Deleted");
+            var result = _repository.GetById(id);
+            if (result != null)
+            {
+                _repository.Delete(id);
+                return Ok($"Content Creator with id {id} cannot be found");
+            }
+            return NotFound($"The id {id} cannot be found ");
         }
     }
 }
