+++ SG.Web/Controllers/UploadController.cs
@@ -26,7 +26,7 @@ namespace SG.Web.Controllers
         {
             return Ok(_repository.GetAll());
         }
-        [HttpGet("get-by-id")]
+        [HttpGet("get-by-id/{id}")]
         public IActionResult GetById(Guid id)
         {
             var result = _repository.GetById(id);
@@ -36,18 +36,36 @@ namespace SG.Web.Controllers
             }
             return NotFound("Not Found");
         }
-        [HttpPatch("update")]
-        public IActionResult Update(UploadModel model)
+        [HttpPatch("update/{id}")]
+        public IActionResult Update(UploadModel model, Guid id)
         {
+            var result = _repository.GetById(id);
+            if (result != null)
+            {
+                result.Summary = String.IsNullOrEmpty(model.Summary) ? result.Summary : model.Summary;
+                result.DateModified = DateTime.Now;
+                result.Department = String.IsNullOrEmpty(model.Department) ? result.Department : model.Department;
 
-            _repository.Update(model);
-            return Ok(model);
+                result.Duration = String.IsNullOrEmpty(model.Duration.ToString()) ? result.Duration : model.Duration;
+                result.ContentCreatorId = String.IsNullOrEmpty(model.ContentCreatorId.ToString()) ? result.ContentCreatorId : model.ContentCreatorId;
+                result.Content = String.IsNullOrEmpty(model.Content) ? result.Content : model.Content;
+                _repository.Update(result);
+                return Ok(result);
+
+            }
+            return NotFound($"The id {id} cannot be found ");
+
         }
-        [HttpDelete("delete")]
+        [HttpDelete("delete/{id}")]
         public IActionResult Delete(Guid id)
         {
-            _repository.Delete(id);
-            return Ok("Deleted");
+            var result = _repository.GetById(id);
+            if (result != null)
+            {
+                _repository.Delete(id);
+                return Ok($"Content Creator with id {id} cannot be found");
+            }
+            return NotFound($"The id {id} cannot be found ");
         }
     }
 }
